Add support for parsing markdown
